add new test and fixed requirements.txt and also code correction
