--- hrm/department_app/models/user_model.py
@@ -1,24 +1,0 @@
-"""User model"""
-from django.dispatch import receiver
-from django.urls import reverse
-from django.core.mail import send_mail
-from django_rest_passwordreset.signals import reset_password_token_created
-
-
-
-@receiver(reset_password_token_created)
-def password_reset_token_created(reset_password_token, *args, **kwargs):
-    """Reset passwrod with lib"""
-    email_plaintext_message = "{}?token={}".format(
-        reverse('password_reset:reset-password-request'), reset_password_token.key)
-
-    send_mail(
-        # title:
-        "Password Reset for {title}".format(title="Some website title"),
-        # message:
-        email_plaintext_message,
-        # from:
-        "noreply@somehost.local",
-        # to:
-        [reset_password_token.user.email]
-    )
