+++ requirements.txt
@@ -2,10 +2,11 @@ Django>=3.1
 django-money
 psycopg2-binary
 psycopg2
-django-rest-framework
 djangorestframework 
 gunicorn 
 PyJWT==2.1.0
 pytz
 django-rest-passwordreset==1.1.0
 django-rest-knox==4.1.0
+coveralls
+python-dateutil
